Rename webgpu/emscripten_webgpu.h to emscripten/webgpu_html5.h
